requirements.txt
Flask
torch
realesrgan
Pillow